refactor(broker): step 3/6 — extract internal.ts and threshold-oracle.ts
Step 3 of the directory split. Extracts two cohesive modules:

  cost-ledger/replay-check/
  ├── internal.ts (135 LOC)
  │     - DB row interfaces (CostEventBatchRow, AgentDayDbRow,
  │       TaskDbRow, BudgetDbRow, HighestLsnRow,
  │       ThresholdCrossingDbRow)
  │     - BATCH_SIZE constant
  │     - Key encoders (agentDayKey, splitAgentDayKey, crossingKey)
  │     - Helpers (arraysEqual, eventTypeToKind, parseStoredThresholds)
  └── threshold-oracle.ts (164 LOC)
        - ExpectedCrossing / ComputeExpectedCrossingsArgs types
        - computeExpectedCrossings + visitCandidates + evalBudget
        - isOracleApplicable + oracleObservedFor
        - crossesThresholdBigInt

`internal.ts` is the private shared helpers module — nothing here is
part of the package's public surface. The DB row interfaces sit next
to the SQL queries that produce them (in the orchestrator file) but
are imported by the comparators downstream.

`threshold-oracle.ts` makes the §15.A drift oracle a standalone
testable unit. The wide-typed `ComputeExpectedCrossingsArgs`
(ReadonlySet / ReadonlyMap on the scope indexes) accommodates the
hostile-Proxy wrappers the perf-regression tests use.

replay-check.ts drops from 1150 LOC to 919 LOC.

All 249 broker tests pass; typecheck clean.
