+++ packages/broker/src/cost-ledger/replay-check.ts
@@ -50,6 +50,27 @@ import type {
   ReplayedCrossing,
 } from "./replay-check/discrepancy.ts";
 import {
+  type AgentDayDbRow,
+  agentDayKey,
+  arraysEqual,
+  BATCH_SIZE,
+  type BudgetDbRow,
+  type CostEventBatchRow,
+  crossingKey,
+  eventTypeToKind,
+  type HighestLsnRow,
+  parseStoredThresholds,
+  splitAgentDayKey,
+  type TaskDbRow,
+  type ThresholdCrossingDbRow,
+} from "./replay-check/internal.ts";
+import {
+  type ComputeExpectedCrossingsArgs,
+  computeExpectedCrossings,
+  crossesThresholdBigInt,
+  type ExpectedCrossing,
+} from "./replay-check/threshold-oracle.ts";
+import {
   flagUnsafeAccumulator,
   MAX_BUDGET_LIMIT_MICRO_USD_BIG,
   MAX_SAFE_INTEGER_BIG,
@@ -60,43 +81,6 @@ export type {
   ReplayDiscrepancy,
   ReplayedBudget,
 } from "./replay-check/discrepancy.ts";
-
-interface CostEventBatchRow {
-  readonly lsn: number;
-  readonly type: string;
-  readonly payload: Buffer;
-}
-
-interface AgentDayDbRow {
-  readonly agentSlug: string;
-  readonly dayUtc: string;
-  // Aggregate totals are read with `.safeIntegers(true)` so a hostile
-  // INTEGER value in the projection (past `Number.MAX_SAFE_INTEGER`)
-  // doesn't silently round in the diagnostic — the "diagnostic of last
-  // resort" preserves exact bytes from the row.
-  readonly totalMicroUsd: bigint;
-}
-
-interface TaskDbRow {
-  readonly taskId: string;
-  readonly totalMicroUsd: bigint;
-}
-
-interface BudgetDbRow {
-  readonly budgetId: string;
-  readonly scope: string;
-  readonly subjectId: string | null;
-  readonly limitMicroUsd: number;
-  readonly thresholdsBps: string;
-  readonly setAtLsn: number;
-  readonly tombstoned: number;
-}
-
-interface HighestLsnRow {
-  readonly lsn: number;
-}
-
-const BATCH_SIZE = 1_000;
 
 export function runReplayCheck(db: Database.Database): ReplayCheckReport {
   const readBatchStmt = db.prepare<[number, number], CostEventBatchRow>(
@@ -402,22 +386,6 @@ export function runReplayCheck(db: Database.Database): ReplayCheckReport {
     };
   });
   return txn.deferred();
-}
-
-interface ThresholdCrossingDbRow {
-  readonly budgetId: string;
-  readonly budgetSetLsn: number;
-  readonly thresholdBps: number;
-  readonly crossedAtLsn: number;
-  // Decimal-string form — see the `CAST(observed_micro_usd AS TEXT)`
-  // comment on listCrossingsStmt. Compare via bigint widening or
-  // string equality; never `Number(...)`.
-  readonly observedMicroUsd: string;
-  readonly limitMicroUsd: number;
-}
-
-function crossingKey(budgetId: string, budgetSetLsn: number, thresholdBps: number): string {
-  return `${budgetId}|${budgetSetLsn}|${thresholdBps}`;
 }
 
 // Internal test seam. The end-to-end path that exercises
@@ -862,206 +830,6 @@ function compareBudgets(
       out.push({ kind: "budget_row_ghost", budgetId: budgetId as BudgetId });
     }
   }
-}
-
-function agentDayKey(agentSlug: string, dayUtc: string): string {
-  // Pipe separator: agentSlug is constrained by `AgentSlug` brand
-  // (lowercase alnum + underscore) and dayUtc is `YYYY-MM-DD`; neither
-  // contains `|`, so a key-collision attack via the agent_slug field is
-  // structurally impossible.
-  return `${agentSlug}|${dayUtc}`;
-}
-
-function splitAgentDayKey(key: string): { readonly agentSlug: string; readonly dayUtc: string } {
-  const idx = key.indexOf("|");
-  if (idx === -1) {
-    throw new Error(`replay-check: malformed agent-day key ${key}`);
-  }
-  return { agentSlug: key.slice(0, idx), dayUtc: key.slice(idx + 1) };
-}
-
-function parseStoredThresholds(raw: string): readonly number[] | { readonly error: string } {
-  let parsed: unknown;
-  try {
-    parsed = JSON.parse(raw);
-  } catch (err) {
-    return { error: err instanceof Error ? err.message : String(err) };
-  }
-  if (!Array.isArray(parsed)) {
-    return { error: "thresholds_bps is not an array" };
-  }
-  // Mirror the protocol's `BudgetSetAuditPayload.thresholdsBps`
-  // validation: each entry must be a positive safe-integer ≤ 10_000.
-  // Explicit `Number.isSafeInteger` + bounds rather than `typeof
-  // === "number"` so values like `Infinity` (which `JSON.parse` can
-  // produce from `"[1e999]"`) are rejected here — they would later
-  // JSON-serialize as `null` and the corruption signal would be
-  // lost by the time it reaches on-call.
-  for (const n of parsed) {
-    if (typeof n !== "number" || !Number.isSafeInteger(n) || n <= 0 || n > 10_000) {
-      return {
-        error: `thresholds_bps contains invalid entry ${String(n)} (expected positive safe integer ≤ 10000)`,
-      };
-    }
-  }
-  return parsed as readonly number[];
-}
-
-function arraysEqual(a: readonly number[], b: readonly number[]): boolean {
-  if (a.length !== b.length) return false;
-  for (let i = 0; i < a.length; i += 1) {
-    if (a[i] !== b[i]) return false;
-  }
-  return true;
-}
-
-function eventTypeToKind(type: string): AuditEventKind | "other" {
-  if (type === "cost.event") return "cost_event";
-  if (type === "cost.budget.set") return "budget_set";
-  if (type === "cost.budget.threshold.crossed") return "budget_threshold_crossed";
-  return "other";
-}
-
-interface ExpectedCrossing {
-  readonly budgetId: string;
-  readonly budgetSetLsn: number;
-  readonly thresholdBps: number;
-  readonly crossedAtLsn: number;
-  /**
-   * Cumulative oracle-observed spend, as `bigint` to preserve precision
-   * across the unbounded accumulator range. Emitted to discrepancies
-   * via `.toString()` (decimal) to honor the `MicroUsd` brand: cumulative
-   * lifetime can exceed `MAX_BUDGET_LIMIT_MICRO_USD` and would forge the
-   * brand if cast back to `number`.
-   */
-  readonly observedMicroUsd: bigint;
-  readonly limitMicroUsd: number;
-  /** `crossedAt` derived from the triggering cost event, as ms-since-epoch. */
-  readonly crossedAtMs: number;
-}
-
-interface ComputeExpectedCrossingsArgs {
-  readonly costEventLsn: number;
-  readonly costEventOccurredAtMs: number;
-  readonly agentSlug: string;
-  readonly taskId: string | undefined;
-  readonly budgets: ReadonlyMap<string, ReplayedBudget>;
-  readonly globalBudgetIds: ReadonlySet<string>;
-  readonly agentBudgetIds: ReadonlyMap<string, ReadonlySet<string>>;
-  readonly taskBudgetIds: ReadonlyMap<string, ReadonlySet<string>>;
-  /** BigInt accumulators preserve precision past `Number.MAX_SAFE_INTEGER`. */
-  readonly globalLifetime: bigint;
-  readonly agentLifetime: ReadonlyMap<string, bigint>;
-  readonly taskLifetime: ReadonlyMap<string, bigint>;
-  readonly out: Map<string, ExpectedCrossing>;
-}
-
-// Independent threshold-crossing oracle for a single cost_event. Mirrors
-// the reactor's logic (`processCostEventForCrossings` in reactor.ts) but
-// derives `observed` from oracle-maintained lifetime trackers rather than
-// reading `cost_by_agent` / `cost_by_task` from disk. The crossing math
-// (`crossesThresholdBigInt`) MUST stay byte-identical with the reactor's
-// `crossesThreshold`; if they diverge, this oracle becomes a tautology.
-function computeExpectedCrossings(args: ComputeExpectedCrossingsArgs): void {
-  // Visit each scope index directly. The replay loop's
-  // `replaceBudgetInIndex` keeps a budget in exactly one scope/subject
-  // bucket at a time, so the three iterations below are disjoint and
-  // no dedupe is required. This avoids one transient Set allocation
-  // and O(events × globalBudgetIds.size) hash inserts per cost event.
-  // The disjointness invariant is asserted by the perf-regression test
-  // in cost-ledger.spec.ts: a hostile Proxy on `args.budgets` blocks
-  // universe iteration, and hostile Proxies on the agent/task scope
-  // maps block `.entries()`/`.values()` regressions that would scan
-  // subjects rather than `.get(slug)` directly.
-  visitCandidates(args, args.globalBudgetIds, evalBudget);
-  visitCandidates(args, args.agentBudgetIds.get(args.agentSlug), evalBudget);
-  if (args.taskId !== undefined) {
-    visitCandidates(args, args.taskBudgetIds.get(args.taskId), evalBudget);
-  }
-}
-
-function visitCandidates(
-  args: ComputeExpectedCrossingsArgs,
-  budgetIds: ReadonlySet<string> | undefined,
-  visit: (args: ComputeExpectedCrossingsArgs, budgetId: string, budget: ReplayedBudget) => void,
-): void {
-  if (budgetIds === undefined) return;
-  for (const budgetId of budgetIds) {
-    const budget = args.budgets.get(budgetId);
-    if (budget === undefined) continue;
-    visit(args, budgetId, budget);
-  }
-}
-
-function evalBudget(
-  args: ComputeExpectedCrossingsArgs,
-  budgetId: string,
-  budget: ReplayedBudget,
-): void {
-  // Defensive guards retained from the original pre-#842 oracle. The
-  // index lifecycle (replaceBudgetInIndex) keeps tombstones out of the
-  // scope sets and `isOracleApplicable` should always be true on a
-  // candidate pulled from its own scope index. Note: these guards only
-  // catch STALE/EXTRA entries — they cannot catch the more dangerous
-  // failure mode where a live, applicable budget is missing from all
-  // candidate indexes (an under-emission, masked here as silent).
-  // The lifecycle tests cover that path directly.
-  if (budget.tombstoned) return;
-  if (!isOracleApplicable(budget, args.agentSlug, args.taskId)) return;
-  const observed = oracleObservedFor(budget, args);
-  if (observed === 0n && budget.limitMicroUsd === 0) return;
-  for (const thresholdBps of budget.thresholdsBps) {
-    const key = crossingKey(budgetId, budget.setAtLsn, thresholdBps);
-    if (args.out.has(key)) continue;
-    if (!crossesThresholdBigInt(observed, budget.limitMicroUsd, thresholdBps)) continue;
-    args.out.set(key, {
-      budgetId,
-      budgetSetLsn: budget.setAtLsn,
-      thresholdBps,
-      // Keep cumulative `observed` as bigint internally. Discrepancy
-      // emission stringifies via `.toString()` to honor the
-      // `MicroUsd` brand contract on the wire.
-      crossedAtLsn: args.costEventLsn,
-      observedMicroUsd: observed,
-      limitMicroUsd: budget.limitMicroUsd,
-      crossedAtMs: args.costEventOccurredAtMs,
-    });
-  }
-}
-
-function isOracleApplicable(
-  budget: ReplayedBudget,
-  agentSlug: string,
-  taskId: string | undefined,
-): boolean {
-  if (budget.scope === "global") return true;
-  if (budget.scope === "agent") return budget.subjectId === agentSlug;
-  if (budget.scope === "task") return taskId !== undefined && budget.subjectId === taskId;
-  return false;
-}
-
-function oracleObservedFor(budget: ReplayedBudget, args: ComputeExpectedCrossingsArgs): bigint {
-  if (budget.scope === "global") return args.globalLifetime;
-  if (budget.scope === "agent") return args.agentLifetime.get(args.agentSlug) ?? 0n;
-  // task scope — caller already gated on taskId !== undefined.
-  if (args.taskId === undefined) return 0n;
-  return args.taskLifetime.get(args.taskId) ?? 0n;
-}
-
-/**
- * Integer threshold test in BigInt — byte-identical to the reactor's
- * `crossesThreshold` (see reactor.ts:313). The operand bounds make the
- * intermediate product reach 1e16, above Number.MAX_SAFE_INTEGER, so
- * Number-multiplication would silently truncate; BigInt is exact across
- * the full documented range. `observed` is a bigint accumulator (see
- * `unsafe_lifetime_accumulator` for the precision rationale); `limit`
- * is the per-budget cap and stays well within safe-integer range, so
- * widening it inside the function is safe. If you change the reactor's
- * math, change this too — otherwise the oracle becomes a tautology.
- */
-function crossesThresholdBigInt(observed: bigint, limit: number, thresholdBps: number): boolean {
-  if (limit === 0) return false;
-  return observed * 10_000n >= BigInt(limit) * BigInt(thresholdBps);
 }
 
 function compareExpectedAndLoggedCrossings(
